[refactor] modify component name & rename for typescript
