+++ src/store/index.js
@@ -1,9 +1,6 @@
 import { createStore } from "vuex";
 
-// import admin from './admin';
-// import adminProduct from "./adminProduct";
 import adminOrder from "./adminOrder";
-// import adminCoupon from "./adminCoupon";
 import client from "./client";
 import clientProduct from "./clientProduct";
 import clientCart from "./clientCart";
@@ -42,10 +39,7 @@ export default createStore({
     },
   },
   modules: {
-    // admin,
-    // adminProduct,
     adminOrder,
-    // adminCoupon,
     client,
     clientProduct,
     clientCart,
